# build/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Status": {},
			"%24Result.DefaultSelection_Prisma.%24ContactPayload_": {
				"properties": {
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"position": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"companyId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"updated_at",
					"created_at",
					"status",
					"position",
					"phone",
					"email",
					"lastName",
					"firstName",
					"companyId",
					"id"
				],
				"type": "object"
			},
			"Contact": {
				"$ref": "#/components/schemas/%24Result.DefaultSelection_Prisma.%24ContactPayload_",
				"description": "Model Contact"
			},
			"ContactInput": {
				"properties": {
					"companyId": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"position": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					}
				},
				"required": [
					"companyId",
					"firstName",
					"lastName",
					"email",
					"phone",
					"position",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ContactInput_": {
				"properties": {
					"companyId": {
						"type": "number",
						"format": "double"
					},
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"position": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"active",
							"inactive"
						]
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "prismatsoa",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "v.santana_0"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/contacts": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Contact"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Contact"
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ContactInput"
							}
						}
					}
				}
			}
		},
		"/contacts/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Contact"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Contact"
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ContactInput_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Contact"
								}
							}
						}
					}
				},
				"tags": [
					"Contact"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}